The database is now created using schemas.
The database structure can now be modified in the db.schemas file.
Changes also include the use of a new table for keeping metadata about
the language, such as schema version, code, or autonym.

New functions are also provided in db.table to handle SQL queries
without writing them by hand.

LLENGUA_DB_VERSION has been moved to a config module, so it can be
imported from the whole project.

The database will use a renovated structure, which merges the phonetic
and ortographic inventory, this is yet to be reflected.